requirements.txt
beautifulsoup4==4.12.3
fire==0.6.0
gradio==4.39.0
ipython==8.12.3
langchain_anthropic==0.1.23
langchain_community==0.2.16
langchain_core==0.2.38
langchain_experimental==0.0.65
langchain_openai==0.1.20
langgraph==0.1.19
Markdown==3.4.1
networkx==3.3
nltk==3.8.1
numpy==1.26.4
python-dotenv==1.0.1
PyYAML==6.0.1
Requests==2.32.3
tavily_python==0.3.3
xmltodict==0.13.0
arxiv==2.1.3
wikipedia==1.4.0
duckduckgo-search==6.2.11
pypandoc==1.13
